Resolved shellcheck warnings
Some checks regarding shell incompatibility were ignored as changing
shell could have unforeseen consequences.

The file "configure" will be addressed separately due to it's size.

+++ c-sdk/ldap/build/compver.sh
@@ -47,18 +47,20 @@ COMP_ROOT=$1
 COMP_VERSION=$2
 COMP_VERSION_FILE=${COMP_ROOT}/Version
 COMPOBJDIR=$3
+# Unused, but leave as changing args could break other tools
+# shellcheck disable=SC2034
 MCOM_ROOT=$4
-MODULE=$5		# Module which needs this component
-COMP_RELEASE=$6		# Component release dir
-COMP_NAME=$7		# component name (e.g. ldapsdk, rouge)
-COMP_SUBDIRS=$8		# subdirs to ftp over
-TEST_FILE=$9		# to test if ftp was successful
+MODULE=$5       # Module which needs this component
+COMP_RELEASE=$6     # Component release dir
+COMP_NAME=$7        # component name (e.g. ldapsdk, rouge)
+COMP_SUBDIRS=$8     # subdirs to ftp over
+TEST_FILE=$9        # to test if ftp was successful
 
-if test -r ${COMP_VERSION_FILE}; then \
-  CUR_VERSION=`cat ${COMP_VERSION_FILE}`; \
+if test -r "${COMP_VERSION_FILE}"; then \
+  CUR_VERSION=$(cat "${COMP_VERSION_FILE}"); \
 
   if test "${CUR_VERSION}" = "${COMP_VERSION}"; then \
-    if test -d ${COMP_ROOT}/${COMPOBJDIR}; then \
+    if test -d "${COMP_ROOT}"/"${COMPOBJDIR}"; then \
       exit 0; \
     fi; \
   fi; \
@@ -70,18 +72,18 @@ echo "The ${COMP_NAME} client libraries are missing.  "
 echo ""
 echo "Attempting to download..."
 
-rm -rf ${COMP_ROOT}/${COMPOBJDIR} ${COMP_VERSION_FILE}
-mkdir -p ${COMP_ROOT}/${COMPOBJDIR}
+rm -rf "${COMP_ROOT:?}"/"${COMPOBJDIR}" "${COMP_VERSION_FILE}"
+mkdir -p "${COMP_ROOT}"/"${COMPOBJDIR}"
 
-sh ../../build/nsftp.sh ${COMP_NAME}/${COMP_VERSION}/${COMPOBJDIR} ${COMP_ROOT}/${COMPOBJDIR}
+sh ../../build/nsftp.sh "${COMP_NAME}"/"${COMP_VERSION}"/"${COMPOBJDIR}" "${COMP_ROOT}"/"${COMPOBJDIR}"
 
 for d in ${COMP_SUBDIRS}; do \
-  mkdir -p ${COMP_ROOT}/${COMPOBJDIR}/${d}; \
-  sh ../../build/nsftp.sh ${COMP_NAME}/${COMP_VERSION}/${COMPOBJDIR}/${d} ${COMP_ROOT}/${COMPOBJDIR}/${d}
+  mkdir -p "${COMP_ROOT}"/"${COMPOBJDIR}"/"${d}"; \
+  sh ../../build/nsftp.sh "${COMP_NAME}"/"${COMP_VERSION}"/"${COMPOBJDIR}"/"${d}" "${COMP_ROOT}"/"${COMPOBJDIR}"/"${d}"
 done
 
-if test -f ${TEST_FILE}; then \
-    echo "${COMP_VERSION}" > ${COMP_VERSION_FILE}; \
+if test -f "${TEST_FILE}"; then \
+    echo "${COMP_VERSION}" > "${COMP_VERSION_FILE}"; \
     echo "************************ SUCCESS! ************************"; \
 else \
     echo ""; \
